requirements.txt
flask
pandas
numpy
scikit-learn
scipy
joblib
matplotlib
seaborn
jupyter
pytest
flake8
black
isort
gunicorn
liac-arff
openpyxl